refactor: MarketingCollections section
